奖惩记录增删改查 (#17)
* 模糊查询

* 模糊查询

* 模糊查询

* 调整个人查询

* 调整个人查询

* 调整个人查询无法出现类别类型

* 调整个人查询无法出现类别类型

* 筛选查询

* 定义对象

* 管理员查找所用用户加上角色

* jwt: token有效期两周，便于调试

* 改了那个

* 那个record的id

* 无

* 1

* 1

* 调整获取userId的方式

* gai

---------

+++ src/main/java/top/suyiiyii/sims/controller/RecordController.java
@@ -2,14 +2,18 @@ package top.suyiiyii.sims.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import lombok.Data;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import top.suyiiyii.sims.common.AuthAccess;
+import top.suyiiyii.sims.common.JwtInterceptor;
 import top.suyiiyii.sims.common.Result;
 import top.suyiiyii.sims.dto.CommonResponse;
 import top.suyiiyii.sims.dto.RecordDto;
 import top.suyiiyii.sims.entity.Record;
+
 import top.suyiiyii.sims.service.CategoryService;
 import top.suyiiyii.sims.service.RecordService;
 import top.suyiiyii.sims.service.RoleService;
@@ -17,6 +21,7 @@ import top.suyiiyii.sims.service.UserService;
 import top.suyiiyii.sims.utils.JwtUtils;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 
 @RestController
@@ -25,7 +30,7 @@ RecordController {
     @Autowired
     RecordService recordService;
     @Autowired
-    UserService UserService;
+    UserService userService;
     @Autowired
     RoleService roleService;
     @Autowired
@@ -42,8 +47,8 @@ RecordController {
         List<Record> records = recordService.getAllRecords(page, size);
         List<RecordDto> recordDtos = new ArrayList<>();
         for (Record record : records) {
-
             RecordDto recordDto = modelMapper.map(record, RecordDto.class);
+            recordDto.setId(record.getId());
             recordDto.setCategoryName(categoryService.getCategoryName(record.getCategoryId()));
             recordDto.setSubCategoryName(categoryService.getsubCategoryName(record.getCategoryId()));
             recordDtos.add(recordDto);
@@ -57,29 +62,27 @@ RecordController {
     public Result<List<RecordDto>> record(@RequestParam(defaultValue = "0") int page,
                                           @RequestParam(defaultValue = "10") int size,
                                           HttpServletRequest request) {
-        String token = (String) request.getAttribute("token");
-        String userId = JwtUtils.extractUserId(token);
+        String userId = String.valueOf(JwtInterceptor.getUserIdFromReq(request));
         List<RecordDto> recordDtos = new ArrayList<>();
-
         List<Record> records = recordService.getMyAllRecords(page, size, userId);
         for (Record record : records) {
-
             RecordDto recordDto = modelMapper.map(record, RecordDto.class);
-
+            recordDto.setId(record.getId());
             recordDto.setCategoryName(categoryService.getCategoryName(record.getCategoryId()));
             recordDto.setSubCategoryName(categoryService.getsubCategoryName(record.getCategoryId()));
-
             recordDtos.add(recordDto);
         }
         return Result.success(recordDtos);
-
     }
-
     @AuthAccess(allowRoles = {"admin"})
     @Operation(summary = "更新单个奖惩记录")
     @PutMapping("/admin/record/{id}")
     public Result<CommonResponse> adminUpdateRecord(@PathVariable Integer id, @RequestBody RecordDto recordDto) {
         Record record = modelMapper.map(recordDto, Record.class);
+        Integer i = recordService.IsRecord(id);
+        if(i==null) {
+            throw new RuntimeException("该记录不存在");
+        }
         recordService.updateRecord(record, id);
         return Result.msg("修改成功");
     }
@@ -88,6 +91,10 @@ RecordController {
     @Operation(summary = "删除单个奖惩记录")
     @DeleteMapping("/admin/record/{id}")
     public Result<CommonResponse> adminDeleteRecord(@PathVariable Integer id) {
+        Integer i = recordService.IsRecord(id);
+        if(i==null) {
+            throw new RuntimeException("该记录不存在");
+        }
         recordService.deleteRecord(id);
         return Result.msg("删除成功");
     }
@@ -97,15 +104,143 @@ RecordController {
     @Operation(summary = "添加奖惩记录")
     @PostMapping("/admin/record")
     public Result<CommonResponse> adminAddRecord(@RequestBody RecordDto recordDto) {
-        Integer categoryId = categoryService.getIdBySubCategoryName(recordDto.getSubCategoryName());
-
-        Record record = modelMapper.map(recordDto, Record.class);
-        if (categoryId == null) {
-            Result.error("请选择奖惩类别，以及类型");
+//CategoryName不是奖励或者惩罚
+        if (!recordDto.getCategoryName().equals("奖励")
+                && !recordDto.getCategoryName().equals("惩罚")) {
+            return Result.error("请选择正确奖惩类别");
         }
-        record.setCategoryId(categoryId);
-        recordService.addRecord(record);
+        if (recordDto.getSubCategoryName().isEmpty()) {
+            return Result.error("请输入奖惩类型");
+        }
+
+        recordService.addRecord(recordDto);
         return Result.msg("添加成功");
     }
+   @AuthAccess(allowRoles = {"admin"})
+    @Operation(summary = "模糊查询奖惩记录")
+    @GetMapping("/admin/likeRecords")
+    public Result<List<RecordDto>> searchRecords(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size,
+            SearchRequest searchRequest) {
 
+        Integer s1;
+        List<Integer> studentIds = new ArrayList<>();
+        List<Record> records=new ArrayList<>();
+       Integer studentId = searchRequest.getStudentId();
+        if(studentId!=null) {
+            studentIds.add(studentId);
+        }
+       String roleName = searchRequest.getRoleName();
+        if(roleName!="") {
+            //rolename查用户id
+            Integer userId = roleService.getIdByrolename(roleName);
+         //     用户id查记录
+          s1 = userService.getStudentIdByUserId(userId);
+            studentIds.add(s1);
+        }
+       String username = searchRequest.getUsername();
+        if(username!="") {
+            //username查用户StudentId
+            s1= roleService.getStudentIdByUsername(username);
+            studentIds.add(s1);
+        }
+        for (Integer Sid : studentIds) {
+            records.addAll(recordService.getRecordsLike(page,size,Sid,searchRequest.getUserGroup(),searchRequest.getGrade()));
+        }
+        List<RecordDto> RecordDtos = new ArrayList<>();
+        for (Record record : records) {
+            RecordDto RecordDto = modelMapper.map(record, RecordDto.class);
+            RecordDto.setCategoryName(categoryService.getCategoryName(record.getCategoryId()));
+            RecordDto.setSubCategoryName(categoryService.getsubCategoryName(record.getCategoryId()));
+            RecordDtos.add(RecordDto);
+        }
+        return Result.success(RecordDtos);
+    }
+    @AuthAccess(allowRoles = {"admin"})
+    @Operation(summary = "筛选查询奖惩记录")
+    @GetMapping("/admin/screenRecords")
+    public Result<List<RecordDto>> screenRecords(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size,
+            String categoryName) {
+        List<Integer> studentIds = new ArrayList<>();
+        //CategoryName不是奖励或者惩罚
+        if (!categoryName.equals("奖励")
+                && !categoryName.equals("惩罚")) {
+            return Result.error("请选择正确奖惩类别");
+        }
+        List<Integer> idByCategoryName = categoryService.getIdByCategoryName(categoryName);
+        for (Integer i : idByCategoryName) {
+            List<Integer> sid = recordService.getSidByCategoryId(i);
+            studentIds.addAll(sid);
+        }
+        List<Record> records=new ArrayList<>();
+        HashSet<Integer> studentIds1= new HashSet<>(studentIds);
+        for (Integer Sid : studentIds1) {
+            if(Sid!=null){
+                records.addAll(recordService.getRecordsById(page,size,Sid));
+            }
+        }
+        List<RecordDto> RecordDtos = new ArrayList<>();
+        for (Record record : records) {
+            RecordDto RecordDto = modelMapper.map(record, RecordDto.class);
+            RecordDto.setCategoryName(categoryService.getCategoryName(record.getCategoryId()));
+            RecordDto.setSubCategoryName(categoryService.getsubCategoryName(record.getCategoryId()));
+            RecordDtos.add(RecordDto);
+        }
+        return Result.success(RecordDtos);
+    }
+
+    @AuthAccess(allowRoles = {"user","admin"})
+    @Operation(summary = "用户筛选查询奖惩记录")
+    @GetMapping("/screenRecords")
+    public Result<List<RecordDto>> screenRecords1(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size,
+            String categoryName,HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        String token = (String) session.getAttribute("token");
+        String userId = JwtUtils.extractUserId(token);
+        if (userId==null){
+            throw new RuntimeException("请先登录");
+        }
+        List<Integer> studentIds = new ArrayList<>();
+        //CategoryName不是奖励或者惩罚
+        if (!categoryName.equals("奖励")
+                && !categoryName.equals("惩罚")) {
+            return Result.error("请选择正确奖惩类别");
+        }
+        List<Integer> idByCategoryName = categoryService.getIdByCategoryName(categoryName);
+        for (Integer i : idByCategoryName) {
+            List<Integer> sid = recordService.getSidByCategoryId(i);
+            if(sid!=null) {
+                studentIds.addAll(sid);
+            }
+        }
+        List<Record> records=new ArrayList<>();
+        HashSet<Integer> studentIds1= new HashSet<>(studentIds);
+        for (Integer Sid : studentIds1) {
+            Integer studentId1 =userService.getStudentIdByUserId(Integer.valueOf(userId));
+            if (studentId1!= null && studentId1.equals(Sid)) {
+                records.addAll(recordService.getRecordsById(page, size, Sid));
+            }
+        }
+        List<RecordDto> RecordDtos = new ArrayList<>();
+        for (Record record : records) {
+            RecordDto RecordDto = modelMapper.map(record, RecordDto.class);
+            RecordDto.setCategoryName(categoryService.getCategoryName(record.getCategoryId()));
+            RecordDto.setSubCategoryName(categoryService.getsubCategoryName(record.getCategoryId()));
+            RecordDtos.add(RecordDto);
+        }
+        return Result.success(RecordDtos);
+    }
+    @Data
+    public static class SearchRequest {
+        private String username;
+        private Integer studentId;
+        private String grade;
+        private String userGroup;
+        private String roleName;
+    }
 }
